+++ app/utils/utils.py
@@ -3,13 +3,15 @@ import os
 import random
 import re
 import socket
+import subprocess
 import sys
 import time
 import unicodedata
+from io import BufferedReader
 from pathlib import Path
 from shutil import which
 from subprocess import PIPE, Popen, check_call
-from typing import Any, Dict, List, Tuple, Union
+from typing import IO, Any, Dict, List, Tuple, Union
 from urllib.parse import urlparse
 
 import click
@@ -20,7 +22,7 @@ from progress.spinner import PixelSpinner
 from progressbar import ETA, Bar, Counter, ProgressBar, Timer
 from stringcolor import bold
 
-from .config import BASE_PATH, ENV_MODE, LOGGING_LEVEL, PROJECT_NAME
+from .config import BASE_PATH, DEBUG, ENV_MODE, LOGGING_LEVEL, PROJECT_NAME
 from .defaultLogBanner import log_runBanner
 
 # ------------------------------------------------------------------------------
@@ -51,6 +53,7 @@ class Context:
         self.disable_split_project: Union[bool, None] = None
         self.disable_split_host: Union[bool, None] = None
         self.print_only_mode: Union[bool, None] = None
+        self.terminal_read_mode: bool = False
 
 
 pass_context = click.make_pass_decorator(Context, ensure=True)
@@ -63,6 +66,16 @@ pass_context = click.make_pass_decorator(Context, ensure=True)
 
 
 class Utils:
+
+    runner_init_count = 1
+    runner_time_check_running = 1
+    runner_text_it_is_running = [
+        "...yep, still running",
+        "...no stress, process still running",
+        "...process is aaalive ;)",
+        "...we current still processing, please wait ... loooong time :P",
+        "...still running bro"
+    ]
 
     # --------------------------------------------------------------------------
     #
@@ -159,10 +172,12 @@ class Utils:
         except Exception:
             pass
 
-    def run_command(self, command_list: List[str] = [], input: Union[str, None] = None,
-                    inner_loop: bool = False) -> Tuple[Union[str, None], Union[str, None]]:
+    def run_command(self, command_list: List[str] = [], input_value: Union[str, None] = None,
+                    ) -> Tuple[Union[str, None], Union[str, None], bool]:
         sub_std_res: Union[str, None] = None
         sub_err_res: Union[str, None] = None
+
+        is_interrupted: bool = False
 
         if not self.ctx.print_only_mode:
             try:
@@ -177,51 +192,16 @@ class Utils:
                     if self.prompt_sudo() != 0:
                         sys.exit(4)
 
-                init_count = 1
-                time_check_running = 1
-                text_it_is_running = [
-                    "...yep, still running",
-                    "...no stress, process still running",
-                    "...process is aaalive ;)",
-                    "...we current still processing, please wait ... loooong time :P",
-                    "...still running bro"
-                ]
-
                 if self.is_tool(command_list[index_to_check]):
-                    if input is None:
+                    if input_value is None:
+                        # , start_new_session=True
                         with Popen(command_list, stdout=PIPE, stderr=PIPE) as sub_p:
-                            time.sleep(time_check_running)
-                            if sub_p.poll() is None:
-                                with PixelSpinner('Processing... ') as spinner:
-                                    while sub_p.poll() is None:
-                                        if init_count % 6 == 0:
-                                            spinner.message = f'{random.choice(text_it_is_running)} '
-                                            init_count = 1
-                                        spinner.next()
-                                        init_count += 1
-                                        time.sleep(time_check_running)
-                            (sub_std, sub_err) = sub_p.communicate()
+                            sub_std, sub_err, is_interrupted = self.subprocess_handler(
+                                sub_p=sub_p, input_value=input_value, command=command_list[index_to_check])
                     else:
                         with Popen(command_list, stdout=PIPE, stderr=PIPE, stdin=PIPE) as sub_p:
-                            if sub_p.stdin is not None and input is not None:
-                                # (sub_std, sub_err) = sub_p.communicate(input=input.encode())
-                                sub_p.stdin.write(input.encode())
-                                sub_p.stdin.close()
-                            time.sleep(time_check_running)
-                            if sub_p.poll() is None:
-                                with PixelSpinner('Processing... ') as spinner:
-                                    while sub_p.poll() is None:
-                                        if init_count % 6 == 0:
-                                            spinner.message = f'{random.choice(text_it_is_running)} '
-                                            init_count = 1
-                                        spinner.next()
-                                        init_count += 1
-                                        time.sleep(time_check_running)
-
-                            if sub_p.stdout is not None:
-                                sub_std = sub_p.stdout.read()
-                            if sub_p.stderr is not None:
-                                sub_err = sub_p.stderr.read()
+                            sub_std, sub_err, is_interrupted = self.subprocess_handler(
+                                sub_p=sub_p, input_value=input_value, command=command_list[index_to_check])
                 else:
                     logging.log(logging.ERROR, f'the command "{command_list[index_to_check]}", did not exist')
                     sub_err = b"MISSING_COMMAND"
@@ -234,12 +214,67 @@ class Utils:
 
             except KeyboardInterrupt as k:
                 logging.log(logging.WARNING, f'process interupted! ({k})')
-                if inner_loop:
-                    raise KeyboardInterrupt
+                is_interrupted = True
             except Exception as e:
                 logging.log(logging.CRITICAL, e, exc_info=True)
+        return (sub_std_res, sub_err_res, is_interrupted)
 
-        return (sub_std_res, sub_err_res)
+    def subprocess_handler(self, sub_p: Popen, input_value: Union[str, None] = None,
+                           command: Union[str, None] = None
+                           ) -> Tuple[Union[bytes, None], Union[bytes, None], bool]:
+        sub_std: Union[bytes, None] = None
+        sub_err: Union[bytes, None] = None
+
+        sub_p_std: Union[IO[bytes], bytes, None] = None
+        sub_p_err: Union[IO[bytes], None] = None
+
+        is_interrupted: bool = False
+
+        try:
+            if sub_p.stdin is not None and input_value is not None:
+                sub_p.stdin.write(input_value.encode())
+                sub_p.stdin.close()
+
+            if sub_p.poll() is None:
+                if not self.ctx.terminal_read_mode or (command is not None and command == 'tee'):
+                    time.sleep(self.runner_time_check_running)
+                    with PixelSpinner('Processing... ') as spinner:
+                        while sub_p.poll() is None:
+                            if self.runner_init_count % 6 == 0:
+                                spinner.message = f'{random.choice(self.runner_text_it_is_running)} '
+                                self.runner_init_count = 1
+                            spinner.next()
+                            self.runner_init_count += 1
+                            time.sleep(self.runner_time_check_running)
+
+                    if sub_p.stdout is not None:
+                        sub_p_std = sub_p.stdout
+                else:
+                    for stdout_line in sub_p.stdout:
+                        if stdout_line is not None and len(stdout_line) > 0:
+                            if sub_p_std is None:
+                                sub_p_std = stdout_line
+                            else:
+                                sub_p_std += stdout_line
+                            logging.log(logging.INFO, stdout_line.decode().replace('\n', ''))
+            if sub_p.stderr is not None:
+                sub_p_err = sub_p.stderr
+        except (SystemExit, KeyboardInterrupt) as e:
+            is_interrupted = True
+            if not self.ctx.terminal_read_mode:
+                if sub_p.stdout is not None:
+                    sub_p_std = sub_p.stdout
+            if sub_p.stderr is not None:
+                sub_p_err = sub_p.stderr
+            sub_p.kill()
+
+        if isinstance(sub_p_std, bytes):
+            sub_std = sub_p_std
+        if isinstance(sub_p_std, BufferedReader):
+            sub_std = sub_p_std.read()
+        if isinstance(sub_p_err, BufferedReader):
+            sub_err = sub_p_err.read()
+        return (sub_std, sub_err, is_interrupted)
 
     def is_tool(self, name: str) -> bool:
         '''
@@ -253,32 +288,37 @@ class Utils:
             default exec function is "run_command" with different
         '''
         cmd_result: Union[str, None] = None
+        is_interrupted: bool = False
         try:
             log_runBanner(msg)
             if len(cmds) <= 1:
                 output = False
             for cmd in cmds:
-                logging.log(logging.NOTICE, ' '.join(cmd))
-                if output:
-                    (cmd_result, std_err) = self.run_command(command_list=cmd, input=cmd_result, inner_loop=True)
-                else:
-                    (cmd_result, std_err) = self.run_command(command_list=cmd, inner_loop=True)
-                if std_err is not None and std_err == "MISSING_COMMAND":
-                    cmd_result = None
-                    break
-                if cmd_result is not None:
-                    if len(cmd_result) > 0:
-                        logging.log(logging.DEBUG, cmd_result)
+                if not is_interrupted or cmd[0] == 'tee':
+                    logging.log(logging.NOTICE, ' '.join(cmd))
+                    if output:
+                        cmd_result, std_err, is_interrupted = self.run_command(
+                            command_list=cmd, input_value=cmd_result)
                     else:
+                        cmd_result, std_err, is_interrupted = self.run_command(command_list=cmd)
+                    if std_err is not None and std_err == "MISSING_COMMAND":
                         cmd_result = None
-                        if output:
-                            logging.log(logging.WARNING, 'no result available to pipe')
-                            break
-            return cmd_result
+                        break
+                    if cmd_result is not None:
+                        if len(cmd_result) > 0:
+                            logging.log(logging.SPAM, cmd_result)
+                        else:
+                            cmd_result = None
+                            if output:
+                                logging.log(logging.WARNING, 'no result available to pipe')
+                                break
         except KeyboardInterrupt as k:
             logging.log(logging.WARNING, f'process interupted! ({k})')
+            raise KeyboardInterrupt
         except Exception as e:
             logging.log(logging.CRITICAL, e, exc_info=True)
+        if is_interrupted and cmd_result is None:
+            raise KeyboardInterrupt
         return cmd_result
 
     # --------------------------------------------------------------------------