+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - update nmap usage
 - improve/refactor code
 - add alpha shell script to install dependencies
+- on subprocess run and kill in mid of process, some process which will piped into tee
+  will save the result which are always proceeded to stdout
 
 ### Changed